refactor: use a general purpose kiln steps
also refactor: move exec to shared_state, it makes more sense there
