+++ internal/acceptance/workflows/scenario/step_funcs_tile_source_code.go
@@ -173,8 +173,6 @@ func iAddACompiledSReleaseSourceToTheKilnfile(ctx context.Context, bucketName st
 		SecretAccessKey: accessKey,
 	})
 
-	setPublishableReleaseSource(ctx, bucketName)
-
 	return saveAsYAML(kfPath, kf)
 }
